bug fixing in ret_val struct

+++ misc/teletest.h
@@ -27,7 +27,7 @@ typedef struct
 typedef struct
 {
     float fval;
-    float ival;
+    int ival;
     int valid;
     DataType type;
 } ReturnValue;
